+++ books/serializers.py
@@ -1,0 +1,23 @@
+from rest_framework import serializers
+from .models import Author, Category, Book
+
+class AuthorSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Author
+        fields = ['id', 'name', 'biography']
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields = ['id', 'name', 'description']
+
+class BookSerializer(serializers.ModelSerializer):
+    author = AuthorSerializer(read_only=True)
+    author_id = serializers.PrimaryKeyRelatedField(queryset=Author.objects.all(), source='author', write_only=True)
+
+    category = CategorySerializer(read_only=True)
+    category_id = serializers.PrimaryKeyRelatedField(queryset=Category.objects.all(), source='category', write_only=True)
+
+    class Meta:
+        model = Book
+        fields = ['id', 'title', 'isbn', 'description', 'availability_status', 'author', 'author_id', 'category', 'category_id', 'created_at', 'updated_at']
