+++ library/serializers.py
@@ -1,7 +1,6 @@
 from rest_framework import serializers
-from books.models import Author, Category, Book
-from members.models import Member
-from borrow.models import BorrowRecord
+from books.models import Author, Category
+
 
 class AuthorSerializer(serializers.ModelSerializer):
     class Meta:
@@ -13,27 +12,3 @@ class CategorySerializer(serializers.ModelSerializer):
         model = Category
         fields = ['id', 'name', 'description']
 
-class BookSerializer(serializers.ModelSerializer):
-    author = AuthorSerializer(read_only=True)
-    author_id = serializers.PrimaryKeyRelatedField(queryset=Author.objects.all(), source='author', write_only=True)
-    category = CategorySerializer(read_only=True)
-    category_id = serializers.PrimaryKeyRelatedField(queryset=Category.objects.all(), source='category', write_only=True)
-
-    class Meta:
-        model = Book
-        fields = ['id', 'title', 'isbn', 'author', 'author_id', 'category', 'category_id', 'availability_status']
-
-class MemberSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Member
-        fields = ['id', 'name', 'email', 'membership_date', 'address', 'phone_number']
-
-class BorrowRecordSerializer(serializers.ModelSerializer):
-    book = BookSerializer(read_only=True)
-    book_id = serializers.PrimaryKeyRelatedField(queryset=Book.objects.filter(availability_status=True), source='book', write_only=True)
-    member = MemberSerializer(read_only=True)
-    member_id = serializers.PrimaryKeyRelatedField(queryset=Member.objects.all(), source='member', write_only=True)
-
-    class Meta:
-        model = BorrowRecord
-        fields = ['id', 'book', 'book_id', 'member', 'member_id', 'borrow_date', 'return_date', 'is_returned']
