+++ members/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import Member
+
+class MemberSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Member
+        fields = ['id', 'name', 'email', 'membership_date', 'address', 'phone_number']
